JAMES-1983 CLI test for re-indexing
